+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Support for the `formsAppId` as an optional second property in the `authService.forgotPassword()` function
+- `formService.deleteCivicPlusHCMSContentItem`
 
 ### Changed